some more moved classes
