+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy/src/com/google/cloud/tools/eclipse/appengine/deploy/util/CloudSdkProcessWrapper.java
@@ -19,8 +19,8 @@ package com.google.cloud.tools.eclipse.appengine.deploy.util;
 import com.google.cloud.tools.appengine.api.deploy.AppEngineDeployment;
 import com.google.cloud.tools.appengine.api.deploy.AppEngineStandardStaging;
 import com.google.cloud.tools.appengine.operations.cloudsdk.AppCfg;
-import com.google.cloud.tools.appengine.operations.cloudsdk.CloudSdk;
-import com.google.cloud.tools.appengine.operations.cloudsdk.CloudSdkNotFoundException;
+import com.google.cloud.tools.appengine.operations.CloudSdk;
+import com.google.cloud.tools.appengine.operations.CloudSdkNotFoundException;
 import com.google.cloud.tools.appengine.operations.cloudsdk.Gcloud;
 import com.google.cloud.tools.appengine.operations.cloudsdk.process.LegacyProcessHandler;
 import com.google.cloud.tools.appengine.operations.cloudsdk.process.ProcessHandler;
