+++ plugins/com.google.cloud.tools.eclipse.appengine.libraries/META-INF/MANIFEST.MF
@@ -18,7 +18,8 @@ Export-Package: com.google.cloud.tools.eclipse.appengine.libraries,
  com.google.cloud.tools.eclipse.appengine.libraries.model,
  com.google.cloud.tools.eclipse.appengine.libraries.persistence,
  com.google.cloud.tools.eclipse.appengine.libraries.repository
-Import-Package: com.google.cloud.tools.appengine.operations.cloudsdk;version="0.7.3",
+Import-Package: com.google.cloud.tools.appengine.operations;version="0.7.3",
+ com.google.cloud.tools.appengine.operations.cloudsdk;version="0.7.3",
  com.google.cloud.tools.eclipse.appengine.ui,
  com.google.cloud.tools.eclipse.usagetracker,
  com.google.cloud.tools.eclipse.util,
