+++ plugins/com.google.cloud.tools.eclipse.test.util/src/com/google/cloud/tools/eclipse/test/util/MockSdkGenerator.java
@@ -27,7 +27,7 @@ import java.nio.file.Path;
 
 /**
  * Utility class to generate a mock Google Cloud SDK installation that passes
- * {@code com.google.cloud.tools.appengine.operations.cloudsdk.CloudSdk#validateCloudSdk()}.
+ * {@code com.google.cloud.tools.appengine.operations.CloudSdk#validateCloudSdk()}.
  */
 public class MockSdkGenerator {
   /**
